- model `displayCredit` as a boolean rather than a string - move `Image` type into Media model and export it to be used by flexible-content
